+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.12.1
+
+* Fixes overflow bug on Table component.
+
 # 0.12.0
 
 ## Minor Improvements